create: bot controller. refractoring: changed config file name
